+++ Editor/Scripts/Drawers/ConditionalAttributeDrawers/EnableFieldDrawer.cs
@@ -17,7 +17,7 @@ namespace EditorAttributes.Editor
 
 			var propertyField = DrawProperty(property);
 
-			UpdateVisualElement(() =>
+			UpdateVisualElement(root, () =>
 			{
 				propertyField.SetEnabled(GetConditionValue(conditionalProperty, enableAttribute, property, errorBox));